requirements.txt
pyacoustid
librosa
numpy
scipy
